fixed minimax

+++ src/AI/Node.java
@@ -7,7 +7,7 @@ public class Node<T> implements GenericTree<T> {
 
     private T data = null;
     private int depth = 0;
-    private int value;
+    private double value;
     private List<Node<T>> children = new ArrayList<>();
     private Node<T> parent = null;
     private boolean visited = false;
@@ -76,9 +76,9 @@ public class Node<T> implements GenericTree<T> {
         this.depth = depth;
     }
 
-    public int getValue() { return value; }
+    public double getValue() { return value; }
 
-    public void setValue(int value) { this.value = value; }
+    public void setValue(double value) { this.value = value; }
 
     public int getRow() {
         int row = -1; //makes sure we throw an error if we have not updated our coordinate
